test: move support functions from topos to test sdk
