+++ crates/topos-test-sdk/src/certificates/mod.rs
@@ -1,4 +1,7 @@
+use std::collections::HashMap;
+
 use topos_core::uci::Certificate;
+use topos_core::uci::SubnetId;
 
 use rstest::*;
 
@@ -9,7 +12,7 @@ use crate::constants::TARGET_SUBNET_ID_1;
 #[fixture]
 pub fn create_certificate_chain(
     #[default(SOURCE_SUBNET_ID_1)] source_subnet: topos_core::uci::SubnetId,
-    #[default(TARGET_SUBNET_ID_1)] target_subnet: topos_core::uci::SubnetId,
+    #[default(&[TARGET_SUBNET_ID_1])] target_subnets: &[topos_core::uci::SubnetId],
     #[default(1)] number: usize,
 ) -> Vec<Certificate> {
     let mut certificates = Vec::new();
@@ -21,7 +24,7 @@ pub fn create_certificate_chain(
             source_subnet,
             Default::default(),
             Default::default(),
-            &[target_subnet],
+            target_subnets,
             0,
             Default::default(),
         )
@@ -31,4 +34,27 @@ pub fn create_certificate_chain(
     }
 
     certificates
+}
+
+/// Generate and assign nb_cert number of certificates to existing subnets
+/// Could be different number of certificates per subnet
+pub fn create_certificate_chains(
+    subnets: &[SubnetId],
+    number_of_certificates_per_subnet: usize,
+) -> HashMap<SubnetId, Vec<Certificate>> {
+    let mut result: HashMap<SubnetId, Vec<Certificate>> = HashMap::new();
+
+    subnets.iter().for_each(|subnet| {
+        let targets = subnets
+            .iter()
+            .filter(|sub| *sub != subnet)
+            .copied()
+            .collect::<Vec<_>>();
+
+        let certs =
+            create_certificate_chain(*subnet, targets.as_ref(), number_of_certificates_per_subnet);
+        result.entry(*subnet).or_insert(certs);
+    });
+
+    result
 }
