+++ crates/topos-test-sdk/src/tce/mod.rs
@@ -1,6 +1,9 @@
+use std::collections::HashMap;
 use std::error::Error;
 use std::future::IntoFuture;
+use std::time::Duration;
 
+use futures::future::join_all;
 use futures::Stream;
 use libp2p::identity::Keypair;
 use libp2p::{Multiaddr, PeerId};
@@ -12,6 +15,7 @@ use tonic::transport::Channel;
 use topos_core::api::tce::v1::{
     api_service_client::ApiServiceClient, console_service_client::ConsoleServiceClient,
 };
+use topos_core::api::tce::v1::{PushPeerListRequest, StatusRequest};
 use topos_core::uci::Certificate;
 use topos_core::uci::SubnetId;
 use topos_p2p::{error::P2PError, Client, Event, Runtime};
@@ -189,4 +193,90 @@ pub async fn start_node(
         connected_subnets: None,
         shutdown_sender,
     }
+}
+
+fn build_peer_config_pool(peer_number: u8) -> Vec<NodeConfig> {
+    (1..=peer_number).map(NodeConfig::from_seed).collect()
+}
+
+pub async fn start_pool(
+    peer_number: u8,
+    correct_sample: usize,
+    g: fn(usize, f32) -> usize,
+) -> HashMap<PeerId, TceContext> {
+    let mut clients = HashMap::new();
+    let peers = build_peer_config_pool(peer_number);
+
+    let mut await_peers = Vec::new();
+    for config in &peers {
+        let mut config = config.clone();
+        config.correct_sample = correct_sample;
+        config.g = g;
+
+        let fut = async {
+            let client = start_node(vec![], config, &peers).await;
+
+            (client.peer_id, client)
+        };
+
+        await_peers.push(fut);
+    }
+
+    for (user_peer_id, client) in join_all(await_peers).await {
+        clients.insert(user_peer_id, client);
+    }
+
+    clients
+}
+
+pub async fn create_network(
+    peer_number: usize,
+    correct_sample: usize,
+) -> HashMap<PeerId, TceContext> {
+    let g = |a, b: f32| ((a as f32) * b).ceil() as usize;
+
+    // List of peers (tce nodes) with their context
+    let mut peers_context = start_pool(peer_number as u8, correct_sample, g).await;
+    let all_peers: Vec<PeerId> = peers_context.keys().cloned().collect();
+
+    // Force TCE nodes to recreate subscriptions and subscribers
+    info!("Trigger the new network view");
+    for (peer_id, client) in peers_context.iter_mut() {
+        let _ = client
+            .console_grpc_client
+            .push_peer_list(PushPeerListRequest {
+                request_id: None,
+                peers: all_peers
+                    .iter()
+                    .filter_map(|key| {
+                        if key == peer_id {
+                            None
+                        } else {
+                            Some(key.to_string())
+                        }
+                    })
+                    .collect::<Vec<_>>(),
+            })
+            .await
+            .expect("Can't send PushPeerListRequest");
+    }
+
+    tokio::time::sleep(Duration::from_secs(2)).await;
+
+    // Waiting for new network view
+    let mut status: Vec<bool> = Vec::new();
+
+    for (_peer_id, client) in peers_context.iter_mut() {
+        let response = client
+            .console_grpc_client
+            .status(StatusRequest {})
+            .await
+            .expect("Can't get status");
+
+        status.push(response.into_inner().has_active_sample);
+    }
+
+    assert!(status.iter().all(|s| *s));
+
+    peers_context
 }
